llm-observability-sdk/examples/genai_service/pyproject.toml
[project]
name = "genai-service"
version = "0.1.0"
description = "FastAPI service demonstrating GenAI observability with dual-backend tracing"
requires-python = ">=3.13"
dependencies = [
    # FastAPI and server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.1",

    # Google ADK and GenAI
    "google-adk>=1.2.0",
    "google-genai>=1.0.0",

    # OpenTelemetry core
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-http>=1.28.0",

    # OpenInference for GenAI tracing
    "openinference-instrumentation-google-adk>=0.1.5",

    # Arize for GenAI trace export
    "arize-otel>=0.6.0",

    # GCS integration for patient records
    "google-cloud-storage>=2.14.0",

    # FHIR parsing
    "fhir.resources>=7.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

# =============================================================================
# Ruff Configuration
# =============================================================================

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# MyPy Configuration
# =============================================================================

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = false  # Relaxed for demo service
ignore_missing_imports = true  # External packages may lack stubs
explicit_package_bases = true  # Avoid duplicate module detection
files = ["main.py", "agents.py", "observability.py"]
